Move `com.velocityctd.proxy.commands` -> `com.velocityctd.proxy.command` to align with velocity naming
