+++ proxy/src/main/java/com/velocityctd/proxy/command/builtin/HubCommand.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.velocityctd.proxy.commands.builtin;
+package com.velocityctd.proxy.command.builtin;
 
 import static java.util.Objects.requireNonNull;
