+++ proxy/src/main/java/com/velocitypowered/proxy/VelocityServer.java
@@ -22,19 +22,19 @@ import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import com.velocityctd.proxy.commands.builtin.AlertCommand;
-import com.velocityctd.proxy.commands.builtin.AlertRawCommand;
-import com.velocityctd.proxy.commands.builtin.FindCommand;
-import com.velocityctd.proxy.commands.builtin.GipCommand;
-import com.velocityctd.proxy.commands.builtin.GkickCommand;
-import com.velocityctd.proxy.commands.builtin.HubCommand;
-import com.velocityctd.proxy.commands.builtin.LeaveQueueCommand;
-import com.velocityctd.proxy.commands.builtin.PingCommand;
-import com.velocityctd.proxy.commands.builtin.PlistCommand;
-import com.velocityctd.proxy.commands.builtin.ProxyAliasCommand;
-import com.velocityctd.proxy.commands.builtin.QueueAdminCommand;
-import com.velocityctd.proxy.commands.builtin.SlashServerCommand;
-import com.velocityctd.proxy.commands.builtin.TransferCommand;
+import com.velocityctd.proxy.command.builtin.AlertCommand;
+import com.velocityctd.proxy.command.builtin.AlertRawCommand;
+import com.velocityctd.proxy.command.builtin.FindCommand;
+import com.velocityctd.proxy.command.builtin.GipCommand;
+import com.velocityctd.proxy.command.builtin.GkickCommand;
+import com.velocityctd.proxy.command.builtin.HubCommand;
+import com.velocityctd.proxy.command.builtin.LeaveQueueCommand;
+import com.velocityctd.proxy.command.builtin.PingCommand;
+import com.velocityctd.proxy.command.builtin.PlistCommand;
+import com.velocityctd.proxy.command.builtin.ProxyAliasCommand;
+import com.velocityctd.proxy.command.builtin.QueueAdminCommand;
+import com.velocityctd.proxy.command.builtin.SlashServerCommand;
+import com.velocityctd.proxy.command.builtin.TransferCommand;
 import com.velocityctd.proxy.queue.RedisVelocityQueueManager;
 import com.velocityctd.proxy.queue.VelocityQueueManager;
 import com.velocityctd.proxy.redis.VelocityRedis;
